+++ spec/fixtures/components/home-page/home-page.component.nodom.spec.expected.ts
@@ -1,0 +1,48 @@
+import { TestBed } from '@angular/core/testing';
+import { createSpyObj } from 'jest-createspyobj';
+import { HomePageComponent } from './home-page.component';
+import { Router, ActivatedRoute } from '@angular/router';
+import { ConfigData } from "./config-data.class";
+import { Environment, TEST_ENV } from "./environment.class";
+
+describe('HomePageComponent', () => {
+  let component: HomePageComponent;
+  let fakeRouter: jest.Mocked<Router>;
+  let fakeRoute: jest.Mocked<ActivatedRoute>;
+  let fakeEnvironment: Environment;
+  let fakeData: ConfigData;
+
+  beforeEach(async () => {
+    fakeRouter = createSpyObj<Router>(Router, ['navigate']);
+    fakeRoute = createSpyObj<ActivatedRoute>(ActivatedRoute, []);
+    fakeEnvironment = {} as Environment;
+    fakeData = {} as ConfigData;
+
+    await TestBed.configureTestingModule({
+      declarations: [HomePageComponent],
+      providers: [
+        { provide: Router, useFactory: () => fakeRouter },
+        { provide: ActivatedRoute, useFactory: () => fakeRoute },
+        { provide: TEST_ENV, useValue: fakeEnvironment },
+        { provide: ConfigData, useValue: fakeData },
+        HomePageComponent
+      ]
+    });
+  });
+
+  beforeEach(() => {
+    component = TestBed.inject(HomePageComponent);
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+
+  describe('METHOD: ngOnInit', () => {
+    it.skip('NOT IMPLEMENTED: should do something', () => {
+      // TODO implement test
+      // component.ngOnInit();
+    });
+  });
+
+});