+++ src/main.ts
@@ -1,36 +1,37 @@
-import { readFileSync, writeFileSync, readdirSync } from 'fs';
+import {readFileSync, writeFileSync, readdirSync, existsSync} from 'fs';
 import * as ts from 'typescript';
 import { parseSourceFile } from './parse-source-file';
 import { generateUnitTest } from './generate-unit-test';
 import defaultDependencyHandler from './default-dependency-handler';
 import { DependencyHandler } from './model';
 
-export function run(params: string[]) {
-  if (!params.length) {
+export function run(argv: any) {
+  if (!argv._) {
     // tslint:disable-next-line:no-console
     console.error('missing path argument');
     process.exit(1);
   }
 
-  if (params.length > 1 && params[0].indexOf('--require') === 0) {
-    require(params[1]);
-    params = params.slice(2);
+  if (argv.require) {
+    require(argv.require);
   }
 
   const handlers: DependencyHandler[] = [];
-  if (params.length > 1 && params[0].indexOf('--handlers') === 0) {
-    const files = readdirSync(params[1]);
+  if (argv.handlers) {
+    const files = readdirSync(argv.handlers);
     files.forEach((file) => {
-      const value = require(process.cwd() + '/' + params[1] + '/' + file);
+      const value = require(process.cwd() + '/' + argv.handlers + '/' + file);
       handlers.push(value.default || value);
     });
-    params = params.slice(2);
   }
   handlers.push(defaultDependencyHandler);
 
-  const path = params[0];
+  const path = argv._[0];
 
   const specPath = path.substring(0, path.length - 2) + 'spec.ts';
+  if(!argv.overwrite && existsSync(specPath)) {
+    throw `${specPath} already exists`;
+  }
   const sourceCode = readFileSync(path).toString();
 
   const sourceFile = ts.createSourceFile(
@@ -41,7 +42,7 @@ export function run(params: string[]) {
   );
 
   const input = parseSourceFile(sourceFile);
-  const output = generateUnitTest(path, sourceCode, input, handlers);
+  const output = generateUnitTest(path, sourceCode, input, handlers, argv.nodom);
 
   writeFileSync(specPath, output);
 }