+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v1.1.0
+- command line arguments are more flexible (using yargs)
+- "--nodom" will treat component like a service - no compilation
+- existing spec files will not be overwritten unless "--overwrite" is provided
+
 ## v1.0.6
 Add code to skip unimplemented tests by default